refactor: move navigation fns to separate files for better separation of concerns; update imports
