+++ apps/web/cypress/e2e/pages/accounts_modal.pages.js
@@ -1,0 +1,77 @@
+// AccountsModal (All Accounts popup)
+const allAccountsBtn = '[data-testid="all-accounts-btn"]'
+const importBtn = '[data-testid="import-btn"]'
+const accountsList = '[data-testid="accounts-list"]'
+const pinnedAccounts = '[data-testid="pinned-accounts"]'
+const emptyPinnedList = '[data-testid="empty-pinned-list"]'
+const addSafeButton = '[data-testid="add-safe-button"]'
+const bookmarkIcon = '[data-testid="bookmark-icon"]'
+const missingSignatureInfo = '[data-testid="missing-signature-info"]'
+const readOnlyChip = '[data-testid="read-only-chip"]'
+const pendingActivationIcon = '[data-testid="pending-activation-icon"]'
+const pendingActivationChip = '[data-testid="pending-activation-chip"]'
+const safeItemCard = '[data-testid="safe-item-card"]'
+const safeOptionsBtn = '[data-testid="safe-options-btn"]'
+const renameBtn = '[data-testid="rename-btn"]'
+const dropdownContent = '[data-slot="select-content"]'
+
+export function openAccountsModal() {
+  cy.get('[data-testid="space-safes-navigation-block"]').should('be.visible')
+  cy.get('[data-testid="open-safes-icon"]').click()
+  cy.get(dropdownContent).should('be.visible')
+  cy.get(allAccountsBtn).scrollIntoView().should('be.visible').click()
+  cy.get(accountsList).should('be.visible')
+}
+
+export function verifyAccountsListVisible() {
+  cy.get(accountsList).should('be.visible')
+}
+
+export function verifyPinnedAccountsSectionVisible() {
+  cy.get(pinnedAccounts).should('be.visible')
+}
+
+export function verifyEmptyPinnedList() {
+  cy.get(emptyPinnedList).should('be.visible')
+}
+
+export function clickAddSafeButton() {
+  cy.get(addSafeButton).should('be.visible').click()
+}
+
+export function clickBookmarkIconByIndex(index) {
+  cy.get(bookmarkIcon).eq(index).should('be.visible').click()
+}
+
+export function unpinSafeByName(name) {
+  cy.get(accountsList).contains(name).closest(safeItemCard).find(bookmarkIcon).click()
+}
+
+export function verifyMissingSignatureInfo(threshold, owners) {
+  cy.get(missingSignatureInfo).should('be.visible').and('contain.text', `${threshold}/${owners}`)
+}
+
+export function verifyReadOnlyChipVisible() {
+  cy.get(readOnlyChip).should('be.visible')
+}
+
+export function verifyPendingActivationIconVisible() {
+  cy.get(pendingActivationIcon).should('be.visible')
+}
+
+export function clickSafeOptionsBtn(index = 0) {
+  cy.get(safeOptionsBtn).eq(index).should('be.visible').click()
+  cy.get(renameBtn).should('be.visible')
+}
+
+export function clickRenameBtn() {
+  cy.get(renameBtn).should('be.visible').click()
+}
+
+export function verifyAccountsListContains(name) {
+  cy.get(accountsList).should('contain.text', name)
+}
+
+export function clickOnImportBtn() {
+  cy.get(importBtn).scrollIntoView().should('be.visible').click()
+}
