+++ apps/web/cypress/e2e/regression/import_export_data_2.cy.js
@@ -1,6 +1,6 @@
 import * as file from '../pages/import_export.pages.js'
 import * as constants from '../../support/constants.js'
-import * as safeNav from '../pages/safe_navigation.pages.js'
+import * as accountsModal from '../pages/accounts_modal.pages.js'
 import { getSafes, CATEGORIES } from '../../support/safes/safesHandler.js'
 
 let staticSafes = []
@@ -22,11 +22,11 @@ describe('Import Export Data tests 2', { defaultCommandTimeout: 20000 }, () => {
   })
 
   it('Verify that the Import button opens an import modal', () => {
-    safeNav.clickOnImportBtn()
+    accountsModal.clickOnImportBtn()
   })
 
   it('Verify that correctly formatted json file can be uploaded and shows data', () => {
-    safeNav.clickOnImportBtn()
+    accountsModal.clickOnImportBtn()
     file.dragAndDropFile(validJsonPath)
     file.verifyImportMessages()
     file.verifyImportBtnStatus(constants.enabledStates.enabled)
@@ -38,18 +38,18 @@ describe('Import Export Data tests 2', { defaultCommandTimeout: 20000 }, () => {
   })
 
   it('Verify that only json files can be imported', () => {
-    safeNav.clickOnImportBtn()
+    accountsModal.clickOnImportBtn()
     file.dragAndDropFile(invalidJsonPath)
     file.verifyErrorOnUpload()
     file.verifyImportBtnStatus(constants.enabledStates.disabled)
   })
 
   it('Verify that json files with wrong information are rejected', () => {
-    safeNav.clickOnImportBtn()
+    accountsModal.clickOnImportBtn()
     file.dragAndDropFile(invalidJsonPath_3)
     file.verifyUploadErrorMessage(file.importErrorMessages.noImportableData)
     file.clickOnCancelBtn()
-    safeNav.clickOnImportBtn()
+    accountsModal.clickOnImportBtn()
     file.dragAndDropFile(invalidJsonPath_2)
     file.verifyUploadErrorMessage(file.importErrorMessages.noImportableData)
     file.clickOnCancelBtn()
